requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
loguru
google-genai
python-dotenv
beanie
PyJWT
google-auth
google-auth-oauthlib
google-auth-httplib2
